remove BaseuiProvider #103
